Tweak lodox overrides and make target
Make sure the current version Lodox is used and that the project is
cleaned, compiled and that the doc profile is clean before calling

    rebar3 lfe lodox

While this is a bit elaborate, it works, repeatably.

N.B. make docs works too, albeit with redundant cleaning, though I
don't have the right Ruby deps to build slate docs locally.